+++ SVGAPlayer/ViewController.m
@@ -50,6 +50,7 @@ static SVGAParser *parser;
                        @"https://github.com/yyued/SVGA-Samples/blob/master/rose.svga?raw=true",
                        ];
     [UIApplication sharedApplication].networkActivityIndicatorVisible = YES;
+//    parser.enabledMemoryCache = YES;
     [parser parseWithURL:[NSURL URLWithString:items[arc4random() % items.count]]
          completionBlock:^(SVGAVideoEntity * _Nullable videoItem) {
              [UIApplication sharedApplication].networkActivityIndicatorVisible = NO;
@@ -66,10 +67,13 @@ static SVGAParser *parser;
                                                 NSParagraphStyleAttributeName: para,
                                             }];
                  [self.aPlayer setAttributedText:str forKey:@"banner"];
+
                  [self.aPlayer startAnimation];
+                 
+//                 [self.aPlayer startAnimationWithRange:NSMakeRange(10, 25) reverse:YES];
              }
          } failureBlock:nil];
-
+//
 //        [parser parseWithURL:[NSURL URLWithString:@"https://github.com/svga/SVGA-Samples/raw/master_aep/BitmapColorArea1.svga"] completionBlock:^(SVGAVideoEntity * _Nullable videoItem) {
 //            if (videoItem != nil) {
 //                self.aPlayer.videoItem = videoItem;