+++ Source/SVGAParser.m
@@ -142,6 +142,8 @@ static NSOperationQueue *unzipQueue;
                 [videoItem resetAudiosWithProtoObject:protoObject];
                 if (self.enabledMemoryCache) {
                     [videoItem saveCache:cacheKey];
+                } else {
+                    [videoItem saveWeakCache:cacheKey];
                 }
                 if (completionBlock) {
                     [[NSOperationQueue mainQueue] addOperationWithBlock:^{
@@ -168,6 +170,8 @@ static NSOperationQueue *unzipQueue;
                     [videoItem resetSpritesWithJSONObject:JSONObject];
                     if (self.enabledMemoryCache) {
                         [videoItem saveCache:cacheKey];
+                    } else {
+                        [videoItem saveWeakCache:cacheKey];
                     }
                     if (completionBlock) {
                         [[NSOperationQueue mainQueue] addOperationWithBlock:^{
@@ -239,6 +243,8 @@ static NSOperationQueue *unzipQueue;
                 [videoItem resetAudiosWithProtoObject:protoObject];
                 if (self.enabledMemoryCache) {
                     [videoItem saveCache:cacheKey];
+                } else {
+                    [videoItem saveWeakCache:cacheKey];
                 }
                 if (completionBlock) {
                     [[NSOperationQueue mainQueue] addOperationWithBlock:^{
@@ -294,6 +300,8 @@ static NSOperationQueue *unzipQueue;
                                 [videoItem resetSpritesWithProtoObject:protoObject];
                                 if (self.enabledMemoryCache) {
                                     [videoItem saveCache:cacheKey];
+                                } else {
+                                    [videoItem saveWeakCache:cacheKey];
                                 }
                                 if (completionBlock) {
                                     [[NSOperationQueue mainQueue] addOperationWithBlock:^{
@@ -320,6 +328,8 @@ static NSOperationQueue *unzipQueue;
                                     [videoItem resetSpritesWithJSONObject:JSONObject];
                                     if (self.enabledMemoryCache) {
                                         [videoItem saveCache:cacheKey];
+                                    } else {
+                                        [videoItem saveWeakCache:cacheKey];
                                     }
                                     if (completionBlock) {
                                         [[NSOperationQueue mainQueue] addOperationWithBlock:^{