+++ Source/SVGAVideoEntity.m
@@ -31,11 +31,17 @@
 @implementation SVGAVideoEntity
 
 static NSCache *videoCache;
+static NSMapTable * weakCache;
+static dispatch_semaphore_t videoSemaphore;
 
 + (void)load {
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
         videoCache = [[NSCache alloc] init];
+        weakCache = [[NSMapTable alloc] initWithKeyOptions:NSPointerFunctionsStrongMemory
+        valueOptions:NSPointerFunctionsWeakMemory
+            capacity:64];
+        videoSemaphore = dispatch_semaphore_create(1);
     });
 }
 
@@ -83,7 +89,8 @@ static NSCache *videoCache;
             [JSONImages enumerateKeysAndObjectsUsingBlock:^(NSString * _Nonnull key, NSString * _Nonnull obj, BOOL * _Nonnull stop) {
                 if ([obj isKindOfClass:[NSString class]]) {
                     NSString *filePath = [self.cacheDir stringByAppendingFormat:@"/%@.png", obj];
-                    NSData *imageData = [NSData dataWithContentsOfFile:filePath];
+//                    NSData *imageData = [NSData dataWithContentsOfFile:filePath];
+                    NSData *imageData = [NSData dataWithContentsOfFile:filePath options:NSDataReadingMappedIfSafe error:NULL];
                     if (imageData != nil) {
                         UIImage *image = [[UIImage alloc] initWithData:imageData scale:2.0];
                         if (image != nil) {
@@ -153,7 +160,8 @@ static NSCache *videoCache;
                 filePath = [self.cacheDir stringByAppendingFormat:@"/%@", fileName];
             }
             if ([[NSFileManager defaultManager] fileExistsAtPath:filePath]) {
-                NSData *imageData = [NSData dataWithContentsOfFile:filePath];
+//                NSData *imageData = [NSData dataWithContentsOfFile:filePath];
+                NSData *imageData = [NSData dataWithContentsOfFile:filePath options:NSDataReadingMappedIfSafe error:NULL];
                 if (imageData != nil) {
                     UIImage *image = [[UIImage alloc] initWithData:imageData scale:2.0];
                     if (image != nil) {
@@ -203,11 +211,26 @@ static NSCache *videoCache;
 }
 
 + (SVGAVideoEntity *)readCache:(NSString *)cacheKey {
-    return [videoCache objectForKey:cacheKey];
+    dispatch_semaphore_wait(videoSemaphore, DISPATCH_TIME_FOREVER);
+    SVGAVideoEntity * object = [videoCache objectForKey:cacheKey];
+    if (!object) {
+        object = [weakCache objectForKey:cacheKey];
+    }
+    dispatch_semaphore_signal(videoSemaphore);
+
+    return  object;
 }
 
 - (void)saveCache:(NSString *)cacheKey {
+    dispatch_semaphore_wait(videoSemaphore, DISPATCH_TIME_FOREVER);
     [videoCache setObject:self forKey:cacheKey];
+    dispatch_semaphore_signal(videoSemaphore);
+}
+
+- (void)saveWeakCache:(NSString *)cacheKey {
+    dispatch_semaphore_wait(videoSemaphore, DISPATCH_TIME_FOREVER);
+    [weakCache setObject:self forKey:cacheKey];
+    dispatch_semaphore_signal(videoSemaphore);
 }
 
 @end